optimize My Node
