+++ src/config/index.ts
@@ -1,3 +1,5 @@
-export * from './theme';
+export * from "./theme"
 
-export const API_HOST = process.env.REACT_APP_API_HOST ?? 'https://api-worker-devnet.octopus-network.workers.dev';
+export const API_HOST =
+  process.env.REACT_APP_API_HOST ??
+  "https://api-worker-devnet.octopus-network.workers.dev"
