+++ src/views/Appchain/MyNode/index.tsx
@@ -7,35 +7,23 @@ import {
   Heading,
   Button,
   Text,
-  List,
   Icon,
   Flex,
-  SimpleGrid,
   Select,
   Spinner,
-  useClipboard,
   Center,
-  Link,
   HStack,
   Menu,
   MenuButton,
   MenuItem,
   MenuList,
   Input,
-  Tag,
   IconButton,
   useColorModeValue,
   useBoolean,
-  Skeleton,
 } from "@chakra-ui/react"
 
-import {
-  DownloadIcon,
-  DeleteIcon,
-  CheckIcon,
-  CopyIcon,
-  RepeatIcon,
-} from "@chakra-ui/icons"
+import { DeleteIcon } from "@chakra-ui/icons"
 
 import { CloseIcon } from "@chakra-ui/icons"
 import { BsFillTerminalFill } from "react-icons/bs"
@@ -49,8 +37,9 @@ import type { ApiPromise } from "@polkadot/api"
 import { FcGoogle } from "react-icons/fc"
 
 import { InstanceInfoModal } from "./InstanceInfoModal"
-import { AppchainInfo } from "types"
+import { AppchainInfo, CLOUD_VENDOR } from "types"
 import { useWalletSelector } from "components/WalletSelectorContextProvider"
+import NodeBoard from "components/AppChain/NodeBoard"
 
 type MyNodeProps = {
   appchainId: string | undefined
@@ -59,23 +48,13 @@ type MyNodeProps = {
   appchain?: AppchainInfo
 }
 
-const cloudVendorInLocalStorage =
-  window.localStorage.getItem("OCTOPUS_DEPLOYER_CLOUD_VENDOR") || ""
+const cloudVendorInLocalStorage = window.localStorage.getItem(
+  "OCTOPUS_DEPLOYER_CLOUD_VENDOR"
+) as CLOUD_VENDOR
 const accessKeyInLocalStorage =
   window.localStorage.getItem("OCTOPUS_DEPLOYER_ACCESS_KEY") ||
   window.localStorage.getItem("accessKey") ||
   ""
-
-const statesRecord: any = {
-  "0": { label: "Init", color: "blue", state: 0 },
-  "10": { label: "Applying", color: "teal", state: 10 },
-  "11": { label: "Apply Failed", color: "red", state: 11 },
-  "12": { label: "Running", color: "octo-blue", state: 12 },
-  "20": { label: "Destroying", color: "teal", state: 20 },
-  "21": { label: "Destroy Failed", color: "orange", state: 21 },
-  "22": { label: "Destroyed", color: "gray", state: 22 },
-  "30": { label: "Upgrading", color: "green", state: 30 },
-}
 
 const OAUTH_SCOPE =
   "https://www.googleapis.com/auth/cloud-platform.read-only https://www.googleapis.com/auth/compute"
@@ -88,8 +67,8 @@ export const MyNode: React.FC<MyNodeProps> = ({
 }) => {
   const bg = useColorModeValue("white", "#15172c")
 
-  const [cloudVendor, setCloudVendor] = useState<string>(
-    cloudVendorInLocalStorage || "AWS"
+  const [cloudVendor, setCloudVendor] = useState<CLOUD_VENDOR>(
+    cloudVendorInLocalStorage || CLOUD_VENDOR.AWS
   )
   const [accessKey, setAccessKey] = useState<string>(accessKeyInLocalStorage)
   const [node, setNode] = useState<any>()
@@ -97,10 +76,6 @@ export const MyNode: React.FC<MyNodeProps> = ({
   const [isInitializing, setIsInitializing] = useBoolean()
   const [isLoadingNode, setIsLoadingNode] = useBoolean()
   const [isDeploying, setIsDeploying] = useBoolean()
-  const [isDeleting, setIsDeleting] = useBoolean()
-  const [isApplying, setIsApplying] = useBoolean()
-  const [isRefreshing, setIsRefreshing] = useBoolean()
-  const [isDestroying, setIsDestroying] = useBoolean()
   const [isUpgrading, setIsUpgrading] = useBoolean()
 
   const [nodeMetrics, setNodeMetrics] = useState<any>()
@@ -124,13 +99,6 @@ export const MyNode: React.FC<MyNodeProps> = ({
   const { data: deployConfig } = useSWR("deploy-config")
 
   const { accountId } = useWalletSelector()
-  const { hasCopied: hasInstanceCopied, onCopy: onCopyInstance } = useClipboard(
-    node?.instance ? `${node.instance.user}@${node.instance.ip}` : ""
-  )
-
-  const { hasCopied: hasNodeIdCopied, onCopy: onCopyNodeId } = useClipboard(
-    node?.uuid || ""
-  )
 
   useEffect(() => {
     window.gapi.load("client", () => {
@@ -283,61 +251,11 @@ export const MyNode: React.FC<MyNodeProps> = ({
       })
   }
 
-  const onRefresh = () => {
-    setIsRefreshing.on()
-    axios
-      .get(
-        `${API_HOST}/node/${cloudVendor}/${accessKey}/${appchainId}/${accountId}`
-      )
-      .then((res) => res.data)
-      .then((res) => {
-        if (res) {
-          setNode(res)
-        }
-
-        setIsRefreshing.off()
-      })
-  }
-
   const onClearCache = () => {
     window.localStorage.removeItem("OCTOPUS_DEPLOYER_CLOUD_VENDOR")
     window.localStorage.removeItem("OCTOPUS_DEPLOYER_ACCESS_KEY")
     window.localStorage.removeItem("accessKey")
     window.location.reload()
-  }
-
-  const onApplyNode = () => {
-    let secretKey
-
-    if (cloudVendor === "AWS") {
-      secretKey = window.prompt(
-        "Please enter the secret key of your server",
-        ""
-      )
-
-      if (!secretKey) {
-        return
-      }
-    } else {
-      const { access_token } = oauthUser.getAuthResponse()
-      secretKey = access_token
-    }
-
-    setIsApplying.on()
-    axios
-      .put(
-        `${deployConfig.deployApiHost}/tasks/${node?.uuid}`,
-        {
-          action: "apply",
-          secret_key: secretKey,
-        },
-        {
-          headers: { authorization: node?.user },
-        }
-      )
-      .then((res) => {
-        window.location.reload()
-      })
   }
 
   const onUpgradeImage = () => {
@@ -374,51 +292,6 @@ export const MyNode: React.FC<MyNodeProps> = ({
           headers: { authorization: node?.user },
         }
       )
-      .then((res) => {
-        window.location.reload()
-      })
-  }
-
-  const onDestroyNode = () => {
-    let secretKey
-
-    if (cloudVendor === "AWS") {
-      secretKey = window.prompt(
-        "Please enter the secret key of your server",
-        ""
-      )
-
-      if (!secretKey) {
-        return
-      }
-    } else {
-      const { access_token } = oauthUser.getAuthResponse()
-      secretKey = access_token
-    }
-
-    setIsDestroying.on()
-    axios
-      .put(
-        `${deployConfig.deployApiHost}/tasks/${node?.uuid}`,
-        {
-          action: "destroy",
-          secret_key: secretKey,
-        },
-        {
-          headers: { authorization: node?.user },
-        }
-      )
-      .then((res) => {
-        window.location.reload()
-      })
-  }
-
-  const onDeleteNode = () => {
-    setIsDeleting.on()
-    axios
-      .delete(`${deployConfig.deployApiHost}/tasks/${node?.uuid}`, {
-        headers: { authorization: node?.user },
-      })
       .then((res) => {
         window.location.reload()
       })
@@ -527,142 +400,14 @@ export const MyNode: React.FC<MyNodeProps> = ({
             />
           </Center>
         ) : node ? (
-          <Box mt={3}>
-            <List spacing={3}>
-              <Flex justifyContent="space-between">
-                <Text variant="gray" fontSize="sm">
-                  Status
-                </Text>
-                <Skeleton isLoaded={!isRefreshing}>
-                  <Tag colorScheme={statesRecord[node.state]?.color} size="sm">
-                    {statesRecord[node.state]?.label}
-                  </Tag>
-                </Skeleton>
-              </Flex>
-              <Flex justifyContent="space-between">
-                <Text variant="gray" fontSize="sm">
-                  Node ID
-                </Text>
-                <HStack>
-                  <Text
-                    fontSize="sm"
-                    whiteSpace="nowrap"
-                    w="calc(160px - 30px)"
-                    overflow="hidden"
-                    textOverflow="ellipsis"
-                  >
-                    {node.uuid}
-                  </Text>
-                  <IconButton
-                    aria-label="copy"
-                    onClick={onCopyNodeId}
-                    size="xs"
-                  >
-                    {hasNodeIdCopied ? <CheckIcon /> : <CopyIcon />}
-                  </IconButton>
-                </HStack>
-              </Flex>
-              <Flex justifyContent="space-between">
-                <Text variant="gray" fontSize="sm">
-                  Instance
-                </Text>
-                {node.instance ? (
-                  <HStack>
-                    <Text
-                      fontSize="sm"
-                      whiteSpace="nowrap"
-                      w="calc(160px - 30px)"
-                      overflow="hidden"
-                      textOverflow="ellipsis"
-                    >
-                      {node.instance.user}@{node.instance.ip}
-                    </Text>
-                    <IconButton
-                      aria-label="copy"
-                      onClick={onCopyInstance}
-                      size="xs"
-                    >
-                      {hasInstanceCopied ? <CheckIcon /> : <CopyIcon />}
-                    </IconButton>
-                  </HStack>
-                ) : (
-                  "-"
-                )}
-              </Flex>
-            </List>
-            <Box mt={3}>
-              {node?.state === "0" ? (
-                <SimpleGrid columns={{ base: 1, md: 2 }} spacing={4}>
-                  <Button
-                    colorScheme="octo-blue"
-                    onClick={onApplyNode}
-                    isDisabled={isApplying}
-                    isLoading={isApplying}
-                  >
-                    Apply
-                  </Button>
-                  <Button
-                    onClick={onDeleteNode}
-                    isDisabled={isDeleting}
-                    isLoading={isDeleting}
-                  >
-                    <Icon as={DeleteIcon} mr={2} boxSize={3} /> Delete
-                  </Button>
-                </SimpleGrid>
-              ) : node?.state === "10" || node?.state === "20" ? (
-                <SimpleGrid columns={1}>
-                  <Button
-                    onClick={onRefresh}
-                    isDisabled={isRefreshing}
-                    isLoading={isRefreshing}
-                  >
-                    <RepeatIcon mr={1} /> Refresh
-                  </Button>
-                </SimpleGrid>
-              ) : node?.state === "11" || node?.state === "21" ? (
-                <SimpleGrid columns={1}>
-                  <Button
-                    colorScheme="red"
-                    onClick={onDestroyNode}
-                    isDisabled={isDestroying}
-                    isLoading={isDestroying}
-                  >
-                    <Icon as={DeleteIcon} mr={2} boxSize={3} /> Destroy
-                  </Button>
-                </SimpleGrid>
-              ) : node?.state === "12" ? (
-                <SimpleGrid columns={{ base: 1, md: 2 }} spacing={4}>
-                  <Button as={Link} isExternal href={node.instance.ssh_key}>
-                    <Icon as={DownloadIcon} mr={2} boxSize={3} /> RSA
-                  </Button>
-                  <Button
-                    colorScheme="red"
-                    onClick={onDestroyNode}
-                    isDisabled={isDestroying}
-                    isLoading={isDestroying}
-                  >
-                    <Icon as={DeleteIcon} mr={2} boxSize={3} /> Destroy
-                  </Button>
-                </SimpleGrid>
-              ) : node?.state === "22" ? (
-                <SimpleGrid columns={1}>
-                  <Button
-                    onClick={onDeleteNode}
-                    isDisabled={isDeleting}
-                    isLoading={isDeleting}
-                  >
-                    <Icon as={DeleteIcon} mr={2} boxSize={3} /> Delete
-                  </Button>
-                </SimpleGrid>
-              ) : node?.state === "30" ? (
-                <SimpleGrid columns={1}>
-                  <Button as={Link} isExternal href={node.instance.ssh_key}>
-                    <Icon as={DownloadIcon} mr={2} boxSize={3} /> RSA
-                  </Button>
-                </SimpleGrid>
-              ) : null}
-            </Box>
-          </Box>
+          <NodeBoard
+            node={node}
+            appchainId={appchainId}
+            cloudVendor={cloudVendor}
+            setNode={setNode}
+            deployAccessKey={accessKey}
+            deployConfig={deployConfig}
+          />
         ) : accessKey ? (
           <>
             <Flex minH="120px" justifyContent="center" flexDirection="column">
@@ -737,7 +482,9 @@ export const MyNode: React.FC<MyNodeProps> = ({
                     variant="unstyled"
                     p={2}
                     defaultValue={cloudVendor}
-                    onChange={(e) => setCloudVendor(e.target.value)}
+                    onChange={(e) =>
+                      setCloudVendor(e.target.value as CLOUD_VENDOR)
+                    }
                   >
                     <option value="AWS">AWS</option>
                     <option value="GCP">GCP</option>
