+++ src/views/Appchain/MyStaking/index.tsx
@@ -149,7 +149,7 @@ export const MyStaking: React.FC<MyStakingProps> = ({
         p={6}
         pt={4}
         pb={6}
-        borderRadius="lg"
+        borderBottomRadius="lg"
       >
         {isValidator ? (
           <>
